Revert "Merge changes from develop (mainly for .i files separate)"
This reverts commit 7bede6d4cd896f13eb3b8682355205734a3c534e and
e4889c2b133c1341fd8335547732c456ba629c69

+++ gtsam/nonlinear/NonlinearFactorGraph.cpp
@@ -354,7 +354,7 @@ GaussianFactorGraph::shared_ptr NonlinearFactorGraph::linearize(const Values& li
     _LinearizeOneFactor(*this, linearizationPoint, *linearFG));
 
   // Linearize all non-sendable factors
-  for(size_t i = 0; i < size(); i++) {
+  for(int i = 0; i < size(); i++) {
     auto& factor = (*this)[i];
     if(factor && !(factor->sendable())) {
       (*linearFG)[i] = factor->linearize(linearizationPoint);
